backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "deepagents>=0.5.3",
    "fastapi>=0.136.1",
    "fastmcp>=3.2.4",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-google-genai>=4.2.2",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.1.0",
    "langchain-sandbox>=0.0.3",
    "langgraph>=1.1.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "loguru>=0.7.3",
    "ollama>=0.6.1",
    "pandas>=3.0.2",
    "pydantic-settings>=2.14.0",
    "pypdf>=6.10.2",
    "sqlmodel>=0.0.38",
    "trafilatura>=2.0.0",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "anyio>=4.13.0",
    "httpx>=0.28.1",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
    "vulture>=2.16",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = [
    "E", "F", "W",
    "I",
    "B",
    "UP",
    "SIM",
    "RUF",
    "BLE",
    "TRY",
    "PIE",
    "PL",
]
ignore = ["PLR0913"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004"]
